+++ wm/border.go
@@ -78,17 +78,21 @@ func NewBorder(win fynedesk.Window, icon fyne.Resource, canMaximize bool) *Borde
 	titleBar.title = title
 	titleBar.max = max
 
-	appIcon := canvas.NewImageFromResource(icon)
-	appIcon.SetMinSize(fyne.NewSize(wmTheme.TitleHeight, wmTheme.TitleHeight))
+	if icon != nil {
+		appIcon := &widget.Button{Icon: icon, Importance: widget.LowImportance}
+		appIcon.OnTapped = func() {
+			titleBar.showMenu(appIcon)
+		}
 
-	if buttonPos == "Right" {
-		titleBar.prepend(appIcon)
-		titleBar.prepend(makeFiller(1))
-	} else {
-		titleBar.append(appIcon)
-		titleBar.append(makeFiller(1))
+		if buttonPos == "Right" {
+			titleBar.prepend(appIcon)
+			titleBar.prepend(makeFiller(1))
+		} else {
+			titleBar.append(appIcon)
+			titleBar.append(makeFiller(1))
+		}
+		titleBar.appIcon = appIcon
 	}
-	titleBar.appIcon = appIcon
 
 	return titleBar
 }
@@ -98,7 +102,7 @@ type Border struct {
 	widget.BaseWidget
 	content *fyne.Container
 	focused bool
-	appIcon *canvas.Image
+	appIcon *widget.Button
 	title   *canvas.Text
 	max     *widget.Button
 	win     fynedesk.Window
@@ -123,6 +127,31 @@ func (c *Border) append(obj fyne.CanvasObject) {
 	c.Refresh()
 }
 
+func (c *Border) showMenu(from fyne.CanvasObject) {
+	max := fyne.NewMenuItem("Maximize", func() {
+		if c.win.Maximized() {
+			c.win.Unmaximize()
+		} else {
+			c.win.Maximize()
+		}
+	})
+	if c.win.Maximized() {
+		max.Checked = true
+	}
+	menu := fyne.NewMenu("",
+		fyne.NewMenuItem("Minimize", func() {
+			c.win.Iconify()
+		}),
+		max,
+		fyne.NewMenuItemSeparator(),
+		fyne.NewMenuItem("Close", func() {
+			c.win.Close()
+		}))
+
+	pos := c.win.Position()
+	fynedesk.Instance().ShowMenuAt(menu, pos.Add(from.Position()))
+}
+
 // CreateRenderer creates a new renderer for this border
 //
 // Implements: fyne.Widget
@@ -139,8 +168,7 @@ func (c *Border) SetFocused(focus bool) {
 
 // SetIcon tells the border to change the icon that should be used
 func (c *Border) SetIcon(icon fyne.Resource) {
-	c.appIcon.Resource = icon
-	c.appIcon.Refresh()
+	c.appIcon.SetIcon(icon)
 }
 
 // SetMaximized updates the state of the border maximize indicators and refreshes